requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0
pypdf>=3.0.0
python-dotenv>=1.0.0
ragas>=0.1.0
pydantic>=2.0.0
pdfplumber>=0.7.0
langchain-text-splitters
langchain-chroma
openai